additional auth, tests
- reogranized test
- added additional authorization based on jwt token

+++ src/participant.rs
@@ -6,6 +6,19 @@ use serde::{Deserialize, Serialize};
 
 use crate::{DbState, utils::AppReponse, error::AppError, auth::UserAuth};
 
+#[derive(Deserialize)]
+pub struct CreateParticipant {
+  event: i64,
+  user: i64,
+}
+
+
+#[derive(Serialize)]
+pub struct CreateParticipantResponse {
+  username: String,
+  user: i64,
+}
+
 pub async fn create(
   Json(payload): Json<CreateParticipant>,
   Extension(pool): Extension<DbState>,
@@ -13,7 +26,7 @@ pub async fn create(
 ) -> AppReponse<Json<CreateParticipantResponse>> {
   let CreateParticipant { event, user } = payload;
   if user != auth_userid {
-    return Err(AppError::Unauthorized(format!("cannot make participation for another user, participting user: {}, requester: {}", user, auth_userid)));
+    return Err(AppError::Unauthorized(format!("cannot make participation for another user")));
   }
 
   let selected_user = sqlx::query!(
@@ -48,6 +61,9 @@ pub async fn delete(
   Extension(pool): Extension<DbState>,
   UserAuth(auth_userid): UserAuth,
 ) -> AppReponse<()> {
+  if user_id != auth_userid {
+    return Err(AppError::Unauthorized(format!("cannot remove  participation for another user")));
+  }
   let _ = sqlx::query!(
       r#"
   DELETE FROM participant
@@ -61,15 +77,83 @@ pub async fn delete(
   Ok((StatusCode::NO_CONTENT, ()))
 }
 
-#[derive(Deserialize)]
-pub struct CreateParticipant {
-  event: i64,
-  user: i64,
-}
+#[cfg(test)]
+mod test {
+  use super::*;
+  use serde_json::json;
+  use crate::utils::test::{test_api, setup_with_data};
+  use axum::http;
 
+  mod create {
+    use super::*;
 
-#[derive(Serialize)]
-pub struct CreateParticipantResponse {
-  username: String,
-  user: i64,
+    #[tokio::test]
+    async fn without_auth() {
+      let (app, _) = setup_with_data().await;
+      let body_json = json!({
+        "event": 3,
+        "user": 1,
+      });
+
+      test_api(app, "/participant", http::Method::POST, Some(body_json), None, StatusCode::UNAUTHORIZED, None).await;
+    }
+
+    #[tokio::test]
+    async fn simple() {
+      let (app, pool) = setup_with_data().await;
+      let body_json = json!({
+        "event": 3,
+        "user": 1,
+      });
+      let expected_response = json!({
+        "user": 1,
+        "username": "username1"
+      });
+
+      test_api(app, "/participant", http::Method::POST, Some(body_json), Some(expected_response), StatusCode::CREATED, Some(("1", "username1"))).await;
+
+      let results = sqlx::query!("select * from participant where event = 3 and user = 1")
+        .fetch_all(&pool)
+        .await
+        .unwrap();
+
+      assert_eq!(results.len(), 1);
+    }
+
+    #[tokio::test]
+    async fn for_anothe() {
+      let (app, _) = setup_with_data().await;
+      let body_json = json!({
+        "event": 3,
+        "user": 6,
+      });
+
+      test_api(app, "/participant", http::Method::POST, Some(body_json), None, StatusCode::UNAUTHORIZED, Some(("1", "username1"))).await;
+    }
+  }
+
+  mod delete {
+    use super::*;
+
+    #[tokio::test]
+    async fn simple() {
+      let (app, pool) = setup_with_data().await;
+
+      test_api(app, "/participant/3/1", http::Method::DELETE, None, None, StatusCode::NO_CONTENT, Some(("3", "username3"))).await;
+
+      let results = sqlx::query!("select user, event from participant")
+        .fetch_all(&pool)
+        .await
+        .unwrap();
+
+      assert_eq!(results.len(), 3);
+      println!("{:?}", results);
+      assert_eq!(results[0].user, 2);
+      assert_eq!(results[0].event, 1);
+      assert_eq!(results[1].user, 3);
+      assert_eq!(results[1].event, 2);
+      assert_eq!(results[2].user, 4);
+      assert_eq!(results[2].event, 2);
+    }
+  }
 }
